Support Vagrant libvirt provider
This uses a slightly ugly hack to determine the disk prefix.

It would be excellent if this could instead be set per-provider just
before provisioning the storage.  However, interactions between a
provider block and other Ruby code in the Vagrantfile do not seem well
defined.  Another alternative would be to set:

  :bus => 'scsi',

but then the additional disks would be named sda and sdb, since the
primary disk will still be vda.

Signed-off-by: Martin Schwenke <martin@meltin.net>